refactored error reporting
