+++ common/reporting/sentry_test.go
@@ -1,4 +1,4 @@
-package sentry
+package reporting
 
 import (
 	"errors"
@@ -10,7 +10,7 @@ import (
 
 func TestInitSentry(t *testing.T) {
 	// Call the function to test
-	InitSentry()
+	Init()
 
 	defer sentry.Flush(2 * time.Second)
 	sentry.CaptureException(errors.New("Hello, Sentry!"))
