+++ config/proxy.conf
@@ -1,0 +1,22 @@
+{
+  "addr": "129.80.40.24",
+  "track": "omanyte",
+  "location": {
+    "city": "Ashburn",
+    "country": "United States of America",
+    "countryCode": "US",
+    "latitude": 39.03,
+    "longitude": 77.47
+  },
+  "name": "omanyte-d95807d8-600d-4a27-b09b-4a10031553ce",
+  "port": 80,
+  "protocol": "shadowsocks",
+  "certPem": "LS0tLS1CRUdJTiBDRVJUSUZJQ0FURS0tLS0tCk1JSUIxRENDQVhxZ0F3SUJBZ0lRYjRlczdOTUgza3E3cmVObXZtSG1iakFLQmdncWhrak9QUVFEQWpCSU1SUXcKRWdZRFZRUUtFd3QxYm14bFlYSnVaV1JzZVRFd01DNEdBMVVFQXhNblptOXlaVzFwYzJkcGRtbHVaeUJqWVhSbApjbkJwYkd4aGNteHBhMlVnY0dGdVpHVnlhWHBsTUI0WERUSTBNVEV5TXpFME16VXdObG9YRFRJM01URXlNekUwCk16VXdObG93U0RFVU1CSUdBMVVFQ2hNTGRXNXNaV0Z5Ym1Wa2JIa3hNREF1QmdOVkJBTVRKMlp2Y21WdGFYTm4KYVhacGJtY2dZMkYwWlhKd2FXeHNZWEpzYVd0bElIQmhibVJsY21sNlpUQlpNQk1HQnlxR1NNNDlBZ0VHQ0NxRwpTTTQ5QXdFSEEwSUFCQjJiMTBJcStBbXFSVEQ2dXB3VHJSYjNwN0xSbmlmWDlMY0ZCbkd1UDJsc2N3Vk56YkRhCmZsclpyV0orM2hyVXVVT0tsQU1jeVZsdnV6NDY2ZjhxeXZTalJqQkVNQTRHQTFVZER3RUIvd1FFQXdJRm9EQVQKQmdOVkhTVUVEREFLQmdnckJnRUZCUWNEQVRBTUJnTlZIUk1CQWY4RUFqQUFNQThHQTFVZEVRUUlNQWFIQklGUQpLQmd3Q2dZSUtvWkl6ajBFQXdJRFNBQXdSUUloQUtqbXRVNnFhRUNNMDVjYVc1NHJlUWs4Y1pPQWRmeDVLU3hICnEvd0Z0YURMQWlCNUIxd0Q5RHhONEo2S2E5V21EMWxvZys2VytYUFkrcEZFMDlHeW1JcjJ0QT09Ci0tLS0tRU5EIENFUlRJRklDQVRFLS0tLS0K",
+  "authToken": "0c7a879900b1d56898ba8d8eb1c336620c422688b59f4759a47ac27954809df4",
+  "trusted": true,
+  "connectCfgShadowsocks": {
+    "secret": "z/OGaCkcUplOKTyqHHVtu8bQTUJTOrMMfPL/ICP9ex4=",
+    "cipher": "chacha20-ietf-poly1305",
+    "prefixGenerator": "v1.0 POST /$random_string(5, 10) HTTP/1.1"
+  }
+}
